+++ game-frontend/CHANGELOG.md
@@ -3,6 +3,7 @@
 v0.3.0
 ===========
 * __Feature:__ As a player I would like to be able to explore a planet with my units
+* __Feature:__ As a player, I would like to gather resources from a planet
 
 v0.2.1
 ===========